+++ src/coreclr/scripts/antigen_run.py
@@ -20,8 +20,7 @@ from os import path, walk
 from os.path import getsize
 import os
 from coreclr_arguments import *
-from superpmi_setup import run_command
-from superpmi import TempDir
+from azdo_pipelines_util import run_command, TempDir
 
 parser = argparse.ArgumentParser(description="description")
